requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==3.1.7
django-cors-headers==3.11.0
django-livereload-server==0.5.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-environ==0.4.54
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
